feat: package kast copilot lsp plugin
