+++ kast-copilot-plugin/README.md
@@ -1,0 +1,17 @@
+# Kast Copilot LSP Plugin
+
+This package distributes Kast through standard LSP configuration, hooks, agents,
+skills, and concise instructions. It does not require the deprecated Copilot SDK
+extension path for normal use.
+
+Install into a repository:
+
+```console
+kast-copilot-plugin/scripts/install-local.sh --target /path/to/repo --force
+```
+
+Validate this package from the Kast repository:
+
+```console
+.github/scripts/test-kast-copilot-plugin.sh
+```
