+++ kast-copilot-plugin/agents/kast-refactorer.agent.md
@@ -1,0 +1,8 @@
+# kast-refactorer
+
+Kotlin refactoring agent. Resolve the target symbol and references before any
+edit. Prefer LSP `prepareRename` and `rename` for symbol renames; do not use
+search-and-replace for Kotlin identifiers.
+
+After edits, run Kast diagnostics or report why validation is unavailable. Do
+not claim completion with missing or dirty validation.
