+++ kast-copilot-plugin/hooks/kast-pre-tool-use.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+set -Eeuo pipefail
+
+script_dir="$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)"
+exec python3 "${script_dir}/kast-hook-policy.py" preToolUse "$@"
