+++ kast-copilot-plugin/skills/kast-symbol-investigation/SKILL.md
@@ -1,0 +1,12 @@
+---
+name: kast-symbol-investigation
+description: Use for read-only Kotlin symbol discovery with Kast LSP before falling back to text search.
+---
+
+# Kast Symbol Investigation
+
+1. Search symbols with LSP workspace symbols or document symbols.
+2. Resolve definitions and hover details for exact identity.
+3. Enumerate references and implementations before drawing conclusions.
+4. Use call hierarchy for behavioral impact.
+5. Keep findings compact: symbol, kind, location, reference count, and blockers.
